Conflicts:
	include/cinder/app/App.h
	src/cinder/app/AppImplCocoaBasic.mm
	xcode/cinder.xcodeproj/project.pbxproj

+++ src/cinder/gl/VaoImplSoftware.cpp
@@ -65,7 +65,7 @@ VaoRef createVaoImplSoftware()
 	
 VaoImplSoftware::VaoImplSoftware()
 {
-	mId = ++sIdCounter; // is this adequate? 4billion VAO allocations seems remote; the ID is not really necessary except for debugging
+	mId = (GLuint)++sIdCounter; // is this adequate? 4billion VAO allocations seems remote; the ID is not really necessary except for debugging
 }
 
 VaoImplSoftware::~VaoImplSoftware()
